Fix off-by-one
The `convert()` function *expects* the length to be one byte shorter than
the actual buffer size because it wants to add a NUL in case the buffer
was overrun.

This fixes a segmentation fault when running the tests in a location that
is quite a couple of characters longer than the default `C:\msys32`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>